% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mooncolors}
\alias{mooncolors}
\title{Colors of the Moon}
\format{
A character vector containing 48 color hexcodes.
}
\source{
\url{https://greenflash.photo/greenflash-gallery/greenflash-gallery/portfolio/colors-of-the-moon/}
}
\usage{
mooncolors
}
\description{
A dataset containing 48 different colors of the Moon.
Colors were extracted from photographs taken by Marcella Giulia Pace over a ten year period.
}
\keyword{datasets}
